Created types for Xfrm, Attributes, MiscSelect in sgx-types
